# src/FacilityManagement.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FacilityManagement",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/DigitalAsset/{digitalAssetId}": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Get DigitalAsset by id.",
        "description": "Get DigitalAsset by id.",
        "operationId": "getDigitalAssetById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "digitalAssetId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Delete DigitalAsset.",
        "description": "Delete DigitalAsset.",
        "operationId": "removeDigitalAsset",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "digitalAssetId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveDigitalAssetResponse"
            }
          }
        }
      }
    },
    "/api/DigitalAsset": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Get DigitalAssets.",
        "description": "Get DigitalAssets.",
        "operationId": "getDigitalAssets",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Create DigitalAsset.",
        "description": "Create DigitalAsset.",
        "operationId": "createDigitalAsset",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateDigitalAssetRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateDigitalAssetResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Update DigitalAsset.",
        "description": "Update DigitalAsset.",
        "operationId": "updateDigitalAsset",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateDigitalAssetRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateDigitalAssetResponse"
            }
          }
        }
      }
    },
    "/api/DigitalAsset/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "summary": "Get DigitalAsset Page.",
        "description": "Get DigitalAsset Page.",
        "operationId": "getDigitalAssetsPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetsPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateDigitalAssetRequest": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        }
      }
    },
    "CreateDigitalAssetResponse": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "DigitalAssetDto": {
      "type": "object",
      "properties": {
        "digitalAssetId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "bytes": {
          "format": "byte",
          "type": "string"
        },
        "contentType": {
          "type": "string"
        },
        "height": {
          "format": "float",
          "type": "number"
        },
        "width": {
          "format": "float",
          "type": "number"
        }
      }
    },
    "GetDigitalAssetByIdResponse": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetDigitalAssetsPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DigitalAssetDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetDigitalAssetsResponse": {
      "type": "object",
      "properties": {
        "digitalAssets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DigitalAssetDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "RemoveDigitalAssetResponse": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UpdateDigitalAssetRequest": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        }
      }
    },
    "UpdateDigitalAssetResponse": {
      "type": "object",
      "properties": {
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}