# src/FacilityManagement.App/src/app/@api/models/get-digital-assets-page-response.ts
/* tslint:disable */
import { DigitalAssetDto } from './digital-asset-dto';
export interface GetDigitalAssetsPageResponse {
  entities?: Array<DigitalAssetDto>;
  length?: number;
  validationErrors?: Array<string>;
}